api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.19.0
pydantic==2.6.3
python-multipart==0.0.9
